Refactor all db entities into separate files, refactor common lib for better tree shaking ability
