+++ packages/core/src/lib/rarity.ts
@@ -1,0 +1,58 @@
+import { ElectroError } from 'electrodb';
+import { rarities, type Rarity, type CreateRarity, type UpdateRarity } from 'src/db/rarities';
+import { DBResult } from 'src/types';
+import { getAllCardDesigns } from './design';
+
+export async function getRarityById(args: { rarityId: string }) {
+	const result = await rarities.query.allRarities(args).go();
+	return result.data[0];
+}
+
+export async function getAllRarities() {
+	const result = await rarities.query.allRarities({}).go();
+	return result.data.sort((a, b) => b.defaultCount - a.defaultCount);
+}
+
+export async function createRarity(rarity: CreateRarity): Promise<DBResult<Rarity>> {
+	try {
+		const result = await rarities.create(rarity).go();
+		return { success: true, data: result.data };
+	} catch (err) {
+		if (!(err instanceof ElectroError)) return { success: false, error: `${err}` };
+
+		if (err.code === 4001)
+			// aws error, rarity already exists
+			return {
+				success: false,
+				error: 'Rarity already exists',
+			};
+
+		// default
+		return {
+			success: false,
+			error: err.message,
+		};
+	}
+}
+
+export async function updateRarity(args: UpdateRarity & { rarityId: string }) {
+	const { rarityId, ...rest } = args;
+	const result = await rarities.update({ rarityId }).set(rest).go();
+	return result.data;
+}
+
+export async function deleteRarityById(id: string): Promise<DBResult<Rarity>> {
+	const allDesigns = await getAllCardDesigns();
+	const designsWithRarity = allDesigns.some(
+		(design) => design.rarityDetails?.some((r) => r.rarityId === id)
+	);
+	if (designsWithRarity) {
+		return {
+			success: false,
+			error: 'Cannot delete rarity with existing designs',
+		};
+	}
+
+	const result = await rarities.delete({ rarityId: id }).go();
+	return { success: true, data: result.data };
+}
