+++ packages/core/src/lib/twitch-schemas.ts
